Director folder renaming and clean up
